pyproject.toml
[tool.poetry]
name = "celery-opentelemetry-grpc-issue"
version = "0.1.0"
description = "A quick small replication project of a particular issue."
authors = ["Arden Shackelford <arden@ardens.tech>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8,<3.9"
opentelemetry-sdk = "1.7.1"
opentelemetry-exporter-otlp = "1.8.0"
celery = {extras = ["redis"], version = "5.1.2"}
fastapi = "^0.78.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
requests = "^2.27.1"
grpcio = "1.46.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
